+++ src/main/java/br/com/infocedro/clipper/curation/CurationCandidateEventRepository.java
@@ -1,0 +1,10 @@
+package br.com.infocedro.clipper.curation;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface CurationCandidateEventRepository extends JpaRepository<CurationCandidateEvent, Long> {
+
+    List<CurationCandidateEvent> findByCurationCase_IdOrderByCreatedAtAscIdAsc(Long caseId);
+}